+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [N.N.N] - XXXX-XX-XX
 
 - Move session maintenance to `oh --ctl`
+- Regenerate transcripts inside archived sessions
 
 ## [0.6.0] - 2026-09-18